+++ frontend/src/pages/design/import/component/ReverseDatabase.tsx
@@ -39,7 +39,7 @@ const ReverseDatabase: React.FC<DatabaseReverseProps> = (props) => {
 
   return (<>
     {/*    <ModalForm
-      title={<span>解析已有数据源<span style={{color: "red"}}>（暂时不支持索引解析生成）</span></span>}
+      title={<span>解析已有数据源<span style={{color: "red"}}>（含非主键索引）</span></span>}
       trigger={
         <Button
           key="reverse"
@@ -54,7 +54,7 @@ const ReverseDatabase: React.FC<DatabaseReverseProps> = (props) => {
         return projectDispatch.getSelectedEntity();
       }}
     >*/}
-    <span>解析已有数据源<span style={{color: "red"}}>（暂时不支持索引解析生成）</span></span>
+    <span>解析已有数据源<span style={{color: "red"}}>（含非主键索引）</span></span>
     <StepsForm
       formRef={formRef}
       formProps={{